cli: move the CPU profiling code to its own file
Release note: None
