+++ pkg/cli/cpuprofile.go
@@ -1,0 +1,83 @@
+// Copyright 2020 The Cockroach Authors.
+//
+// Use of this software is governed by the Business Source License
+// included in the file licenses/BSL.txt.
+//
+// As of the Change Date specified in that file, in accordance with
+// the Business Source License, use of this software will be governed
+// by the Apache License, Version 2.0, included in the file
+// licenses/APL.txt.
+
+package cli
+
+import (
+	"context"
+	"os"
+	"path/filepath"
+	"runtime/pprof"
+	"time"
+
+	"github.com/cockroachdb/cockroach/pkg/util/envutil"
+	"github.com/cockroachdb/cockroach/pkg/util/log"
+	"github.com/cockroachdb/cockroach/pkg/util/timeutil"
+)
+
+func initCPUProfile(ctx context.Context, dir string) {
+	const cpuprof = "cpuprof."
+	gcProfiles(dir, cpuprof, maxSizePerProfile)
+
+	cpuProfileInterval := envutil.EnvOrDefaultDuration("COCKROACH_CPUPROF_INTERVAL", -1)
+	if cpuProfileInterval <= 0 {
+		return
+	}
+	if min := time.Second; cpuProfileInterval < min {
+		log.Infof(ctx, "fixing excessively short cpu profiling interval: %s -> %s",
+			cpuProfileInterval, min)
+		cpuProfileInterval = min
+	}
+
+	go func() {
+		defer log.RecoverAndReportPanic(ctx, &serverCfg.Settings.SV)
+
+		ctx := context.Background()
+
+		t := time.NewTicker(cpuProfileInterval)
+		defer t.Stop()
+
+		var currentProfile *os.File
+		defer func() {
+			if currentProfile != nil {
+				pprof.StopCPUProfile()
+				currentProfile.Close()
+			}
+		}()
+
+		for {
+			func() {
+				const format = "2006-01-02T15_04_05.999"
+				suffix := timeutil.Now().Add(cpuProfileInterval).Format(format)
+				f, err := os.Create(filepath.Join(dir, cpuprof+suffix))
+				if err != nil {
+					log.Warningf(ctx, "error creating go cpu file %s", err)
+					return
+				}
+
+				// Stop the current profile if it exists.
+				if currentProfile != nil {
+					pprof.StopCPUProfile()
+					currentProfile.Close()
+					currentProfile = nil
+					gcProfiles(dir, cpuprof, maxSizePerProfile)
+				}
+
+				// Start the new profile.
+				if err := pprof.StartCPUProfile(f); err != nil {
+					log.Warningf(ctx, "unable to start cpu profile: %v", err)
+					f.Close()
+					return
+				}
+				currentProfile = f
+			}()
+		}
+	}()
+}
