+++ pkg/cli/start.go
@@ -208,68 +208,6 @@ func initMemProfile(ctx context.Context, dir string) {
 	}()
 }
 
-func initCPUProfile(ctx context.Context, dir string) {
-	const cpuprof = "cpuprof."
-	gcProfiles(dir, cpuprof, maxSizePerProfile)
-
-	cpuProfileInterval := envutil.EnvOrDefaultDuration("COCKROACH_CPUPROF_INTERVAL", -1)
-	if cpuProfileInterval <= 0 {
-		return
-	}
-	if min := time.Second; cpuProfileInterval < min {
-		log.Infof(ctx, "fixing excessively short cpu profiling interval: %s -> %s",
-			cpuProfileInterval, min)
-		cpuProfileInterval = min
-	}
-
-	go func() {
-		defer log.RecoverAndReportPanic(ctx, &serverCfg.Settings.SV)
-
-		ctx := context.Background()
-
-		t := time.NewTicker(cpuProfileInterval)
-		defer t.Stop()
-
-		var currentProfile *os.File
-		defer func() {
-			if currentProfile != nil {
-				pprof.StopCPUProfile()
-				currentProfile.Close()
-			}
-		}()
-
-		for {
-			func() {
-				const format = "2006-01-02T15_04_05.999"
-				suffix := timeutil.Now().Add(cpuProfileInterval).Format(format)
-				f, err := os.Create(filepath.Join(dir, cpuprof+suffix))
-				if err != nil {
-					log.Warningf(ctx, "error creating go cpu file %s", err)
-					return
-				}
-
-				// Stop the current profile if it exists.
-				if currentProfile != nil {
-					pprof.StopCPUProfile()
-					currentProfile.Close()
-					currentProfile = nil
-					gcProfiles(dir, cpuprof, maxSizePerProfile)
-				}
-
-				// Start the new profile.
-				if err := pprof.StartCPUProfile(f); err != nil {
-					log.Warningf(ctx, "unable to start cpu profile: %v", err)
-					f.Close()
-					return
-				}
-				currentProfile = f
-			}()
-
-			<-t.C
-		}
-	}()
-}
-
 func initBlockProfile() {
 	// Enable the block profile for a sample of mutex and channel operations.
 	// Smaller values provide more accurate profiles but are more
